+++ ext/native/thin3d/thin3d_d3d9.cpp
@@ -572,14 +572,14 @@ public:
 	void DrawUP(const void *vdata, int vertexCount) override;
 	void Clear(int mask, uint32_t colorval, float depthVal, int stencilVal);
 
-	uintptr_t GetNativeObject(NativeObject obj) override {
+	uint64_t GetNativeObject(NativeObject obj) override {
 		switch (obj) {
 		case NativeObject::CONTEXT:
-			return (uintptr_t)d3d_;
+			return (uint64_t)(uintptr_t)d3d_;
 		case NativeObject::DEVICE:
-			return (uintptr_t)device_;
+			return (uint64_t)(uintptr_t)device_;
 		case NativeObject::DEVICE_EX:
-			return (uintptr_t)deviceEx_;
+			return (uint64_t)(uintptr_t)deviceEx_;
 		default:
 			return 0;
 		}
